+++ schoolmgmt/src/main/java/com/diginepal/schoolmgmt/controller/BranchController.java
@@ -3,6 +3,8 @@ package com.diginepal.schoolmgmt.controller;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -15,7 +17,10 @@ import org.springframework.web.bind.annotation.RequestMapping;
 
 import com.diginepal.schoolmgmt.entities.Academicdates;
 import com.diginepal.schoolmgmt.entities.Branch;
+import com.diginepal.schoolmgmt.entities.Company;
 import com.diginepal.schoolmgmt.repo.BranchRepo;
+import com.diginepal.schoolmgmt.response.Response;
+import com.diginepal.schoolmgmt.response.ResponseMessage;
 
 
 @Controller
@@ -26,35 +31,67 @@ public class BranchController {
 		@Autowired
 		BranchRepo branchRepo;
 		
-		@PostMapping (value="/save")
-		public Branch save(@RequestBody Branch branch) {
-			return branchRepo.save(branch);
+		@PostMapping 
+		public ResponseEntity<?> save(@RequestBody Branch branch) {
+			ResponseMessage response=new ResponseMessage();
+			branch=branchRepo.save(branch);
+			if(branch==null) {
+				response=Response.badrequest();
+				return new ResponseEntity<>(response, HttpStatus.BAD_REQUEST);
+			}
+			response=Response.created();
+			return new ResponseEntity<>(response, HttpStatus.CREATED);
 		}
 		
 		@GetMapping (value="/list")
-		public String findAll(Model model){
+		public ResponseEntity<?> findAll(){
+			ResponseMessage response=new ResponseMessage();
 			List<Branch> list=branchRepo.findAll();
-			model.addAttribute("branch", list);
-			return "branch/list";
-		}
-		
-		
+			if(list.isEmpty()) {
+				response=Response.resourcenotfound();
+				return new ResponseEntity<>(response, HttpStatus.NOT_FOUND);
+			}
+			return new ResponseEntity<>(list, HttpStatus.OK);
+		}		
 		@GetMapping (value="/{id}")
-		public Branch findOne (@PathVariable int id){
-			return branchRepo.findById(id).get();
-		}
-
-		@PutMapping(value="/update/{id}")
-		public Branch update(@PathVariable int id,@RequestBody Branch branch) {
-			branch.setId(id);
-			return branchRepo.save(branch);
-	}
-		@DeleteMapping(value="/delete/{id}")
-		public void delete(@PathVariable int id) {
+		public ResponseEntity<?> findOne (@PathVariable int id){
 			Branch branch=branchRepo.findById(id).get();
-			branchRepo.delete(branch);
+			ResponseMessage response=new ResponseMessage();
+			if(branch==null) {
+				response=Response.resourcenotfound();
+				return new ResponseEntity<>(response,HttpStatus.NOT_FOUND);
+			}
+			return new ResponseEntity<>(branch, HttpStatus.OK);
 		}
-
+		@PutMapping(value="/update/{id}")
+		public ResponseEntity<?> update(@PathVariable int id,@RequestBody Branch branch) {
+			ResponseMessage response=new ResponseMessage();
+			Branch search=branchRepo.findById(id).get();
+			if(search==null) {
+				response=Response.resourcenotfound();
+				return new ResponseEntity<>(response,HttpStatus.NOT_FOUND);
+			}
+			else {
+				branch.setId(id);
+				branch=branchRepo.save(branch);
+				response=Response.successful();
+				return new ResponseEntity<>(response, HttpStatus.OK);
+			}
+		}
+		@DeleteMapping(value="/delete/{id}")
+		public ResponseEntity<?> delete(@PathVariable int id) {
+			ResponseMessage response=new ResponseMessage();
+			Branch branch=branchRepo.findById(id).get();
+			if(branch==null) {
+				response=Response.resourcenotfound();
+				return new ResponseEntity<>(response,HttpStatus.NOT_FOUND);
+			}
+			else {
+			branchRepo.delete(branch);
+			response=Response.successful();
+			return new ResponseEntity<>(response,HttpStatus.OK);
+			}
+		}
 	}
