+++ schoolmgmt/src/main/java/com/diginepal/schoolmgmt/controller/EthnicgroupController.java
@@ -3,6 +3,8 @@ package com.diginepal.schoolmgmt.controller;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -14,7 +16,10 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 
 import com.diginepal.schoolmgmt.entities.Ethnicgroup;
+import com.diginepal.schoolmgmt.entities.Faculty;
 import com.diginepal.schoolmgmt.repo.EthnincgroupRepo;
+import com.diginepal.schoolmgmt.response.Response;
+import com.diginepal.schoolmgmt.response.ResponseMessage;
 
 @Controller
 @RequestMapping("ethnicgroup")
@@ -22,33 +27,68 @@ public class EthnicgroupController {
 	@Autowired
 	EthnincgroupRepo ethnicgroupRepo;
 
-	@PostMapping(value = "/save")
-	public Ethnicgroup save(@RequestBody Ethnicgroup ethnicgroup) {
-		return ethnicgroupRepo.save(ethnicgroup);
+	@PostMapping
+	public ResponseEntity<?> save(@RequestBody Ethnicgroup ethnicgroup) {
+		ResponseMessage response=new ResponseMessage();
+		ethnicgroup=ethnicgroupRepo.save(ethnicgroup);
+		if(ethnicgroup==null) {
+			response=Response.badrequest();
+			return new ResponseEntity<>(response, HttpStatus.BAD_REQUEST);
+		}
+		response=Response.created();
+		return new ResponseEntity<>(response, HttpStatus.CREATED);
 	}
-
 	@GetMapping(value="/list")
-	public String findAll(Model model){
+	public ResponseEntity<?> findAll(){
+		ResponseMessage response=new ResponseMessage();
 		List<Ethnicgroup> list=ethnicgroupRepo.findAll();
-		model.addAttribute("ethnicgroup", list);
-		return "ethnicgroup/list";
+		if(list.isEmpty()) {
+			response=Response.resourcenotfound();
+			return new ResponseEntity<>(response, HttpStatus.NOT_FOUND);
+		}
+		return new ResponseEntity<>(list, HttpStatus.OK);
 	}
 
 	@GetMapping(value = "/{id}")
-	public Ethnicgroup findOne(@PathVariable int id) {
-		return ethnicgroupRepo.findById(id).get();
+	public ResponseEntity<?> findOne (@PathVariable int id){
+		Ethnicgroup ethnicgroup=ethnicgroupRepo.findById(id).get();
+		ResponseMessage response=new ResponseMessage();
+		if(ethnicgroup==null) {
+			response=Response.resourcenotfound();
+			return new ResponseEntity<>(response,HttpStatus.NOT_FOUND);
+		}
+		return new ResponseEntity<>(ethnicgroup, HttpStatus.OK);
 	}
+
 
 	@PutMapping(value = "/update/{id}")
-	public Ethnicgroup update(@PathVariable int id, @RequestBody Ethnicgroup ethnicgroup) {
-		ethnicgroup.setId(id);
-		return ethnicgroupRepo.save(ethnicgroup);
+	public ResponseEntity<?> update(@PathVariable int id,@RequestBody Ethnicgroup ethnincgroup) {
+		ResponseMessage response=new ResponseMessage();
+		Ethnicgroup search=ethnicgroupRepo.findById(id).get();
+		if(search==null) {
+			response=Response.resourcenotfound();
+			return new ResponseEntity<>(response,HttpStatus.NOT_FOUND);
+		}
+		else {
+			ethnincgroup.setId(id);
+			ethnincgroup=ethnicgroupRepo.save(ethnincgroup);
+			response=Response.successful();
+			return new ResponseEntity<>(response, HttpStatus.OK);
+		}
 	}
-
 	@DeleteMapping(value = "/delete/{id}")
-	public void delete(@PathVariable int id) {
-		Ethnicgroup ethnicgroup = ethnicgroupRepo.findById(id).get();
+	public ResponseEntity<?> delete(@PathVariable int id) {
+		ResponseMessage response=new ResponseMessage();
+		Ethnicgroup ethnicgroup=ethnicgroupRepo.findById(id).get();
+		if(ethnicgroup==null) {
+			response=Response.resourcenotfound();
+			return new ResponseEntity<>(response,HttpStatus.NOT_FOUND);
+		}
+		else {
 		ethnicgroupRepo.delete(ethnicgroup);
+		response=Response.successful();
+		return new ResponseEntity<>(response,HttpStatus.OK);
+		}
 	}
 
 }
