+++ schoolmgmt/src/main/java/com/diginepal/schoolmgmt/controller/GradeController.java
@@ -3,6 +3,8 @@ package com.diginepal.schoolmgmt.controller;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -14,7 +16,10 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 
 import com.diginepal.schoolmgmt.entities.Grade;
+import com.diginepal.schoolmgmt.entities.Housegroup;
 import com.diginepal.schoolmgmt.repo.GradeRepo;
+import com.diginepal.schoolmgmt.response.Response;
+import com.diginepal.schoolmgmt.response.ResponseMessage;
 
 
 
@@ -25,31 +30,68 @@ public class GradeController {
 	@Autowired
 	GradeRepo gradeRepo;
 	
-	@PostMapping (value="/save")
-	public Grade save(@RequestBody Grade grade) {
-		return gradeRepo.save(grade);
+	@PostMapping 
+	public ResponseEntity<?> save(@RequestBody Grade grade) {
+		ResponseMessage response=new ResponseMessage();
+		grade=gradeRepo.save(grade);
+		if(grade==null) {
+			response=Response.badrequest();
+			return new ResponseEntity<>(response, HttpStatus.BAD_REQUEST);
+		}
+		response=Response.created();
+		return new ResponseEntity<>(response, HttpStatus.CREATED);
 	}
 	
+	
 	@GetMapping(value="/list")
-	public String findAll(Model model){
+	public ResponseEntity<?> findAll(){
+		ResponseMessage response=new ResponseMessage();
 		List<Grade> list=gradeRepo.findAll();
-		model.addAttribute("grade", list);
-		return "grade/list";
+		if(list.isEmpty()) {
+			response=Response.resourcenotfound();
+			return new ResponseEntity<>(response, HttpStatus.NOT_FOUND);
+		}
+		return new ResponseEntity<>(list, HttpStatus.OK);
 	}
 	@GetMapping (value="/{id}")
-	public Grade findOne (@PathVariable int id){
-		return gradeRepo.findById(id).get();
+	public ResponseEntity<?> findOne (@PathVariable int id){
+		Grade grade=gradeRepo.findById(id).get();
+		ResponseMessage response=new ResponseMessage();
+		if(grade==null) {
+			response=Response.resourcenotfound();
+			return new ResponseEntity<>(response,HttpStatus.NOT_FOUND);
+		}
+		return new ResponseEntity<>(grade, HttpStatus.OK);
 	}
-
+	
 	@PutMapping(value="/update/{id}")
-	public Grade update(@PathVariable int id,@RequestBody Grade grade) {
-		grade.setId(id);
-		return gradeRepo.save(grade);
+	public ResponseEntity<?> update(@PathVariable int id,@RequestBody Grade grade) {
+		ResponseMessage response=new ResponseMessage();
+		Grade search=gradeRepo.findById(id).get();
+		if(search==null) {
+			response=Response.resourcenotfound();
+			return new ResponseEntity<>(response,HttpStatus.NOT_FOUND);
+		}
+		else {
+			grade.setId(id);
+			grade=gradeRepo.save(grade);
+			response=Response.successful();
+			return new ResponseEntity<>(response, HttpStatus.OK);
+		}
 }
 	@DeleteMapping(value="/delete/{id}")
-	public void delete(@PathVariable int id) {
+	public ResponseEntity<?> delete(@PathVariable int id) {
+		ResponseMessage response=new ResponseMessage();
 		Grade grade=gradeRepo.findById(id).get();
+		if(grade==null) {
+			response=Response.resourcenotfound();
+			return new ResponseEntity<>(response,HttpStatus.NOT_FOUND);
+		}
+		else {
 		gradeRepo.delete(grade);
+		response=Response.successful();
+		return new ResponseEntity<>(response,HttpStatus.OK);
+		}
 	}
 
 }
